requirements.txt
sentence-transformers 
psycopg2-binary 
faiss-cpu 
pandas
fastapi 
uvicorn
httpx
ollama